+++ bcipy/task/paradigm/matrix/calibration.py
@@ -271,5 +271,8 @@ def init_calibration_display_task(
         stimuli,
         task_bar,
         info,
+        rows=parameters['matrix_rows'],
+        columns=parameters['matrix_columns'],
+        width_pct=parameters['matrix_width'],
         trigger_type=parameters['trigger_type'],
         symbol_set=symbol_set)